XMLParser: Fix MatchToken out of bounds memory access
Unlike strcmp, memcmp won't stop on first '\0' found, and will try to
read out of bounds if the searched string is smaller than the current
token. Unfortunately we can't use strcmp here due to the absence of null
terminator in the token string.
Check for the string length first to make sure both strings are the same
length, and then use memcmp to compare the strings.