Reserve and move expensive vector objects
During Sonic Galactic special stages, a lot of time is spent expanding
the size of arrays/vectors when we already know the max amount
(especially in `GetPolygonPoints()`, which is called a lot).
Polygon2D's constructor is always copying the input vector due to the
explicit `Points = input` assignment.
Reserve space up front to reduce reallocations, and explicitly move the
input vector for Polygon2D's constructor to avoid useless copies.

Ideally this could be done everywhere, but it wouldn't have much of an
impact considering `GetPolygonPoints()` is still the function that
allocates the most even after this commit (~1 million per SS run) and
would bloat the code with `.reserve()` calls.