Dev: Fix viewPerformance display
With `viewPerformance` enabled, the application would just crash because
`Scene::ViewCurrent` was set to -1. Set it to 0 for the default view
instead so that the following code doesn't try to access out of bounds.
After reloading the current scene (F1), the texture font would be
deallocated but `DEBUG_fontSprite` would still be allocated, causing the
font to just disappear. Properly deallocate it on app restart so that it
can be cleanly reallocated.