Implement separate opcode for super invocations
OP_INVOKE reads an additional byte for discerning super calls; this is not
necessary, as most invocations aren't done through super. This splits the
opcode into two, OP_INVOKE and OP_SUPER_INVOKE. This eliminates some additional
checks that needed to be done for determining the correct method to call on the
object. The old opcode still exists as OP_INVOKE_V3, so that bytecode
version 3 and earlier can still be executed correctly.

Additionally, this fixes a bug where super wouldn't correctly use the parent
class for objects that weren't instances (such as doing super within a class.)