Path: Add CONSOLE_FILESYSTEM related API
When set, this enables a much more custom approach to find base path and
pref path.
SDL doesn't necessarily have a sane implementation on consoles, so
having a more fine-grained way of dealing with these paths is simpler.

For Switch, base path is `romfs:/` if using romfs, otherwise it'll
fallback to the cwd. Pref path is hardcoded to `/<GameName>/`.
An optional CONSOLE_BASE_PATH macro can be defined for testing/dev
purposes.